Reorganize Python modules into packages
